backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6

python-dotenv==1.0.1

SQLAlchemy==2.0.31
psycopg2-binary==2.9.10

ariadne==0.23.0

confluent-kafka==2.3.0
kafka-python==2.0.2
six==1.16.0

PyJWT==2.8.0
loguru==0.7.2

prometheus-client==0.20.0

uvicorn==0.30.1
starlette==0.37.2